Documentation/git-log: mention that line-log regex must match in starting revision
When giving a regex as parameter <start> or <end> in
`git log -L <start>,<end>:<file>`, or a function name in
`git log -L :<funcname>:<file>`, the given regex must match in the starting
revision, or else the command exits with a fatal error.

This is not obvious in the documentation, so add a note to that
effect.

Signed-off-by: Philippe Blain <levraiphilippeblain@gmail.com>