parse_branchname_arg(): simplify argument eating
This patch abolishes two pieces of obfuscated code.

First is `if (argc)` condition, which in fact means "more then one arg"
due to `argc++` above.

Second is by far harder to grasp:
    if (!arg0_cant_be_pathspec) {...} else if (opts->accept_pathspec)
that is,
	if (opts->accept_pathspec && arg0_cant_be_pathspec)
which (quite unexpectedly) actually means
    if (opts->accept_pathspec && dash_dash_pos == 1)
and aims to "eat" that `--`.

Make both pieces easier to read by rewriting obfuscated conditions.

With both solved, I could keep argcount++/argv++/argc-- in the very end
of the function, but that was obviously useless code in this case, so I
deleted them as well.

This patch is not expected to change behavior in any way.

Signed-off-by: Alexandr Miloslavskiy <alexandr.miloslavskiy@syntevo.com>