git-gui: add possibility to open currently selected file
...in the default associated app (e.g. in a text editor / IDE).

Many times there's the need to quickly open a source file (the one you're
looking at in Git GUI) in the predefined text editor / IDE. Of course,
the file can be searched for in your preferred file manager or directly
in the text editor, but having the option to directly open the current
file from Git GUI would be just faster. This change enables just that by:
 - Diff header path context menu -> Open;
 - or double-clicking the diff header path.

One "downside" of the approach is that executable files will be run
and not opened for editing.

Signed-off-by: Zoli Szabó <zoli.szabo@gmail.com>