commit: replace rebase/sequence booleans with single pick_state enum
In 116a408, the boolean `rebase_in_progress` was introduced by dscho to
handle instances when cherry-pick and rebase were occuring at the same time.
This created a situation where two independent booleans were being used
to define the state of git at a point in time.

Under his recommendation to follow guidance from Junio, specifically
`https://public-inbox.org/git/xmqqr234i2q0.fsf@gitster-ct.c.googlers.com/`,
it was decided that an `enum` that defines the state of git would be a
more effective path forward.

Tasks completed:
  - Remove multiple booleans `rebase_in_progress` and `sequencer_in_use` and
replaced with a single `pick_state` enum that determines if, when
cherry-picking, we are in a rebase, multi-pick, or single-pick state
  - Converted double `if` statement to `if/else if` prioritizing `REBASE` to
continue to disallow cherry pick in rebase.

Signed-off-by: Ben Curtis <nospam@nowsci.com>