t7063: blindly accept diffs
Assuming that the changes I made in the last commit to drastically
modify how and when and especially how frequently untracked paths are
visited should result in changes to the untracked-cache, this commit
simply updates the t7063 testcases to match what the code now reports.

If this is correct, this commit should be squashed into the previous
one.

It'd be nice if I could get an untracked-cache expert to comment on
this...

Signed-off-by: Elijah Newren <newren@gmail.com>